feat: if a user already exists, the seed operation will not be executed

+++ apps/server/prisma/seed.ts
@@ -3,6 +3,11 @@ import { PrismaClient } from "@prisma/client";
 const prisma = new PrismaClient();
 
 async function main() {
+  const user = await prisma.user.findFirst();
+  if (user) {
+    console.log("User already exists...");
+    return;
+  }
   await prisma.user.deleteMany();
 
   console.log("Seeding...");
